+++ include/survive.h
@@ -1,6 +1,7 @@
 #ifndef _SURVIVE_H
 #define _SURVIVE_H
 
+#include "disambiguator.h"
 #include <stdint.h>
 
 #define SV_FLOAT  		double
@@ -31,10 +32,14 @@ struct SurviveObject
 
 	//Flood info, for calculating which laser is currently sweeping.
 	int8_t oldcode;
+#if 0
 	int16_t total_photos;
 	int32_t last_photo_time;
 	int32_t total_photo_time;
 	int32_t total_pulsecode_time;
+#else
+	disambiguator d;
+#endif
 };
 
 typedef void (*text_feedback_fnptr)( struct SurviveContext * ctx, const char * fault );
