+++ tools/disambiguate/Makefile
@@ -1,4 +1,4 @@
-disambiguate: disambiguate.c
+disambiguate: disambiguate.c ../../src/disambiguator.c
 	gcc -Wall -Wextra -Werror -g -o $@ $^ -lm
 
 test: disambiguate
